sidebar final
