--- src/css/Router.css
@@ -1,28 +1,0 @@
-@import url('./common.css');
-
-.main {
-  display: flex;
-  flex-direction: row;
-}
-
-.side-bar {
-  width: 15%;
-  display: block;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-/* .menu {
-  border: 1px solid black;
-}
-.sub-menu {
-  border: 1px solid black;
-} */
-
-.content {
-  width: 80%;
-}
-
-.vertical {
-  height: 1000px;
-}
