+++ src/css/Sidebar.css
@@ -1,39 +1,60 @@
-/* @import url('./common.css'); */
+.sidebar{
+  display: block;
+  height: 100%;
+  border: none;
+}
+
+.menu-btn{
+  background-color: lightgrey;
+  border: none;
+  font-size: 25px;
+}
 
 .subNav {
   background-color: #060b26;
-  width: 250px;
-  height: 100vh;
   display: flex;
   justify-content: center;
   position: fixed;
-  top: 0;
-  left: -100%;
-  transition: 850ms;
+  top: -100%;
+  transition: 800ms ease all;
 }
 
 .subNav-active {
-  left: 0;
-  transition: 350ms;
+  transition: 800ms ease all;
 }
 
-.sidebar-menu {
-  background-color: #060b26;
-  width: 250px;
-  height: 100vh;
+.big-category-btn {
+  width: 100%;
+  background-color: gray;
+  cursor: pointer;
+  border-radius: 7px;
+  transition: 800ms ease all;
+  border: none;
+  text-align: left;
+}
+
+.big-category-btn:hover {
+  background: black;
+  color: white;
+  cursor: pointer;
+}
+
+.small-category-btn {
+  background: lightgrey;
+  height: 40px;
+  margin: 0.2rem;
   display: flex;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: -100%;
-  transition: 850ms;
+  align-items: center;
+  border: none;
+  cursor: pointer;
+  border-radius: 7px;
+  transition: 800ms ease all;
+  color: black;
+  font-size: 12px;
 }
 
-.sidebar-menu.active {
-  left: 0;
-  transition: 350ms;
-}
-
-.big-menu-btn {
-  background-color: transparent;
+.small-category-btn:hover{
+    background: lightslategray;
+    color: black;
+    cursor: pointer;
 }
