+++ tsconfig.json
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 {
   "compilerOptions": {
     "target": "es6",
@@ -49,55 +48,3 @@
     "craco.config.js"
   ]
 }
-=======
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "esnext",
-    "moduleResolution": "node",
-    "noResolve": false,
-    "noImplicitAny": false,
-    "removeComments": false,
-    "sourceMap": true,
-    "allowJs": true,
-    "jsx": "react-jsx",
-    "allowSyntheticDefaultImports": true,
-    "keyofStringsOnly": true,
-    "skipLibCheck": true,
-    "baseUrl": "./src",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "esModuleInterop": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": false,
-    "noFallthroughCasesInSwitch": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true
-  },
-  "typeRoots": [
-    "node_modules/@types",
-    "src/@type"
-  ],
-  "exclude": [
-    "node_modules",
-    "build",
-    "scripts",
-    "acceptance-tests",
-    "webpack",
-    "jest",
-    "src/setupTests.ts",
-    "./node_modules/**/*"
-  ],
-  "extends": "./tsconfig.paths.json",
-  "include": [
-    "./src/**/*",
-    "@type",
-    "src",
-    "craco.config.js"
-  ]
-}
->>>>>>> 0c7b03e... developing
